refactor(put timeAgo function into util file
